Fix: move scene layer
